requirements.txt
aiogram==3.13.1
flask==3.0.3
sqlalchemy==2.0.35
python-dotenv==1.0.1
werkzeug==3.0.4
gunicorn==23.0.0